Splitting the enums file into separation of concerns
