+++ NearBeach/serializers/kanban_card_serializer.py
@@ -4,7 +4,7 @@ from NearBeach.models import KanbanCard, Project, KanbanColumn, KanbanLevel
 from NearBeach.serializers.enum_serializer import EnumField
 from NearBeach.serializers.kanban_column_serializer import KanbanColumnSerializer
 from NearBeach.serializers.kanban_level_serializer import KanbanLevelSerializer
-from NearBeach.utils.enums import ObjectPriority
+from NearBeach.utils.enums.object_enums import ObjectPriority
 from NearBeach.serializers.tag_serializer import TagSerializer
 from NearBeach.models import Tag, TagAssignment
