+++ NearBeach/utils/enums/request_for_change_enums.py
@@ -2,14 +2,6 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 
-class ObjectPriority(models.IntegerChoices):
-    HIGHEST = 0, _("Highest")
-    HIGH = 1, _("High")
-    NORMAL = 2, _("Normal")
-    LOW = 3, _("Low")
-    LOWEST = 4, _("Lowest")
-
-
 class RequestForChangeImpact(models.IntegerChoices):
     HIGH = 3, _("High"),
     MEDIUM = 2, _("Medium"),
@@ -41,7 +33,7 @@ class RequestForChangeStatus(models.IntegerChoices):
     PAUSED = 7, _("Paused")
     READY_FOR_QA = 8, _("Ready for QA")
     FAILED = 9, _("Failed")
-    
+
 
 class RequestForChangeType(models.IntegerChoices):
     EMERGENCY = 4, _("Emergency")
